api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
neo4j==5.14.0
supabase==2.0.3
python-jose[cryptography]==3.3.0
PyJWT==2.8.1
slowapi==0.1.9
sentry-sdk==1.38.0
httpx==0.25.2
